Refactor(Graduation - BCR) : rename & Add Interface
